Updated the stock_prices simulation ipynb notebook
Created the stock_prices ipynb to run monte carlo simulation on stock prices based on their prices for one year prior to 20 March 2020. This shows where the prices would have been, without the BoC interventions. We can compare this with today's prices to see that the stock prices are double as compared to the simulated prices. This would strengthen our case in the benefits to the banks during this quantitative easing.